requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
bleach==6.3.0
certifi==2026.2.25
charset-normalizer==3.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.25.2
filetype==1.2.0
fonttools==4.62.1
fsspec==2026.2.0
idna==3.7
ImageHash==4.3.2
Jinja2==3.1.6
kaggle==2.0.0
kagglesdk==0.1.16
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
opencv-python==4.13.0.92
opencv-python-headless==4.10.0.84
packaging==26.0
pandas==3.0.1
pi_heif==1.3.0
pillow==12.1.1
pillow-avif-plugin==1.5.5
polars==1.39.3
polars-runtime-32==1.39.3
protobuf==7.34.1
psutil==7.2.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-slugify==8.0.4
PyWavelets==1.9.0
PyYAML==6.0.3
requests==2.33.0
requests-toolbelt==1.0.0
roboflow==1.2.16
scipy==1.17.1
seaborn==0.13.2
setuptools==81.0.0
simsimd==6.5.16
six==1.17.0
stringzilla==4.6.0
sympy==1.14.0
text-unidecode==1.3
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.30
ultralytics-thop==2.0.18
urllib3==2.6.3
webencodings==0.5.1
